requirements.txt
PyYAML==6.0.2
sshkeyboard==2.3.1
